v1

+++ Q11.java
@@ -6,14 +6,14 @@ public class Q11 {
 
 	public static void main(String[] args) {
 
-		// a number of calories and fat grams
 
 		Scanner scan = new Scanner(System.in);
 		System.out.println("Enter the medium ( Air, Water or Steel) : ");
 		String medium = scan.next();
 		System.out.println("Enter the distance in feet : ");
-		int distance = scan.nextInt();
+		double distance = scan.nextDouble();
 
+		
 		switch (medium) {
 		case "Air":
 			System.out.println("The travel time is: " + (distance / 1100));
